Renamed startup.rs to startup_lib.rs.
