+++ src/lib.rs
@@ -1,3 +1,3 @@
 pub mod configuration;
 mod routes;
-pub mod startup;
+pub mod startup_lib;
