+++ src/main.rs
@@ -1,28 +1,28 @@
-use actix_rest_vue_setup::startup;
+use actix_rest_vue_setup::startup_lib;
 use std::env;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
     match args.len() {
-        1 => startup::print_usage(&args[0]),
+        1 => startup_lib::print_usage(&args[0]),
         2 => match args[1].as_ref() {
-            "start" => startup::start(false),
-            "status" => startup::status(),
-            "stop" => startup::stop(false),
-            "restart" => startup::restart(),
-            _ => startup::print_usage(&args[0]),
+            "start" => startup_lib::start(false),
+            "status" => startup_lib::status(),
+            "stop" => startup_lib::stop(false),
+            "restart" => startup_lib::restart(),
+            _ => startup_lib::print_usage(&args[0]),
         },
         3 => {
             if &args[2] == "--force" {
                 match args[1].as_ref() {
-                    "start" => startup::start(true),
-                    "stop" => startup::stop(true),
-                    _ => startup::print_usage(&args[0]),
+                    "start" => startup_lib::start(true),
+                    "stop" => startup_lib::stop(true),
+                    _ => startup_lib::print_usage(&args[0]),
                 }
             } else {
-                startup::print_usage(&args[0])
+                startup_lib::print_usage(&args[0])
             }
         }
-        _ => startup::print_usage(&args[0]),
+        _ => startup_lib::print_usage(&args[0]),
     }
 }
